Support for decimal places, assuming GBP-style (like 12.34)
+ reject additional content beyond the command

+++ src/SupportingProject/BankingTypes.h
@@ -1,4 +1,4 @@
 #pragma once
 
 // Option for variable to store currency
-using Currency = long int;
+using Currency = double;
